+++ CHANGES.rst
@@ -7,6 +7,7 @@ New Features
 -----------------
 
 * #35: Added ``BaseModel.dump_to_csv()``
+* #35: Added ``BaseModel.dump_to_json()``
 
 -----------